backend/pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = "my app"
authors = ["AMKalinin <am1kalinin@gmail.com>"]
readme = "README.md"
packages = [{include = "my_app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
pydantic = {extras = ["email"], version = "^2.1.1"}
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
h5py = "^3.9.0"
python-multipart = "^0.0.6"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
